feat(api): update effects shape, now an array of maps
The effects used to be a single map, keys being the effects id and values the params. This API
seemed a little weird especially for the redux community. Now effects are represented by maps with a
structure really close to a redux action, a map with an effect key to identify the effect to run and
potentially other keys to pass parameters to the handler. This makes returning a batch of effects
simpler, and the api simpler, more familiar to JS users, and easier to type.

BREAKING CHANGE: Effects used to be a map like { effectKey: params, effect2Key: params2 }, now
should become an array of [{ effect: 'effectKey', ...params }, { effect: 'effect2Key', ...params2 }]

1